Add benchmark structure, unified impl file, and updated docs
- Added a dedicated `benchmark` directory for organizing benchmarking scripts and results
- Refactored implementation into a single attention.py file for easier adoption
- Updated README.md with latest benchmark results in Markdown table format
- Updated README.md and setup.py to include required dependencies
