--- attention-comparison-batch2-head4-d32-bwd.csv
@@ -1,7 +1,0 @@
-len,Triton-Attn-Bias (FLOPS),PyTorch-Attn-Bias (FLOPS),PyTorch-Compiled-Attn-Bias (FLOPS),xFormers-Attn-Bias (FLOPS),PyTorch-SDPA (FLOPS)
-256.000000,0.066977,0.057096,0.063406,0.099038,0.143660
-512.000000,0.262998,0.223778,0.237284,0.275190,0.575402
-1024.000000,1.035948,0.897140,0.948383,1.102739,2.272584
-2048.000000,3.842722,2.193728,3.742450,4.245165,8.185317
-4096.000000,10.265305,2.385537,6.371156,12.804008,13.304508
-8192.000000,13.449817,2.472635,6.367071,12.865759,12.901892
