--- attention-comparison-batch2-head4-d32-fwd.csv
@@ -1,7 +1,0 @@
-len,Triton-Attn-Bias (FLOPS),PyTorch-Attn-Bias (FLOPS),PyTorch-Compiled-Attn-Bias (FLOPS),xFormers-Attn-Bias (FLOPS),PyTorch-SDPA (FLOPS)
-256.000000,5.691981,1.630143,2.561522,3.261372,3.224175
-512.000000,13.968662,2.800428,5.423639,7.264354,7.325401
-1024.000000,23.470060,2.543002,5.696219,14.027984,14.096813
-2048.000000,30.674707,2.236218,4.732033,19.294282,19.432868
-4096.000000,33.125575,2.577148,4.070783,18.586986,18.449669
-8192.000000,32.581404,2.816053,4.115240,17.971493,18.066297
