+++ benckmark/attention-comparison-batch2-head4-d32-bwd.csv
@@ -1,0 +1,7 @@
+len,Triton-Attn-Bias (FLOPS),PyTorch-Attn-Bias (FLOPS),PyTorch-Compile-Attn-Bias (FLOPS),xFormers-Attn-Bias (FLOPS),PyTorch-SDPA (FLOPS)
+256.000000,0.073699,0.062677,0.076821,0.077729,0.168817
+512.000000,0.371210,0.242403,0.348554,0.306856,1.007847
+1024.000000,1.257951,1.036895,1.181298,1.738618,6.938603
+2048.000000,4.089672,2.213153,4.083681,4.527996,8.918591
+4096.000000,10.732837,2.395701,6.392552,13.702077,13.685270
+8192.000000,13.440012,2.485934,6.257983,13.159468,13.134775
