+++ benckmark/attention-comparison-batch2-head4-d32-fwd.csv
@@ -1,0 +1,7 @@
+len,Triton-Attn-Bias (FLOPS),PyTorch-Attn-Bias (FLOPS),PyTorch-Compile-Attn-Bias (FLOPS),xFormers-Attn-Bias (FLOPS),PyTorch-SDPA (FLOPS)
+256.000000,5.542667,1.574715,1.943982,3.198991,3.192787
+512.000000,11.284882,2.762875,5.355243,7.125481,7.176634
+1024.000000,23.356817,2.547090,5.691676,13.810132,13.839005
+2048.000000,30.678131,2.247875,4.738464,19.586347,19.654297
+4096.000000,33.154529,2.586121,4.071432,18.912378,18.881439
+8192.000000,32.935548,2.822556,4.118749,18.189223,18.278633
