+++ benckmark/config.py
@@ -8,18 +8,10 @@ Refactor from the original project from [https://github.com/pengzhangzhi]
 
 import triton
 
-configs = [
-    triton.Config({"BLOCK_M": BM, "BLOCK_N": BN}, num_stages=s, num_warps=w)
-    for BM in [32, 64, 128]
-    for BN in [32, 64, 128]
-    for s in ([1, 2])
-    for w in [4, 8]
-]
-
 # Benchmarking configuration
 BATCH, N_HEADS, HEAD_DIM = 2, 4, 32  # Aligned with test_attention_correctness
 
-test_configs = [
+benchmark_configs = [
     triton.testing.Benchmark(
         x_names=["len"],
         x_vals=list(2 ** i for i in range(8, 15)),
@@ -27,14 +19,14 @@ test_configs = [
         line_vals=[
             "triton-attn-bias",
             "torch-attn-bias",
-            "torch-compiled-attn-bias",
+            "torch-compile-attn-bias",
             "xformers-attn-bias",
             "torch-sdpa"
         ],
         line_names=[
             "Triton-Attn-Bias (FLOPS)",
             "PyTorch-Attn-Bias (FLOPS)",
-            "PyTorch-Compiled-Attn-Bias (FLOPS)",
+            "PyTorch-Compile-Attn-Bias (FLOPS)",
             "xFormers-Attn-Bias (FLOPS)",
             "PyTorch-SDPA (FLOPS)"
         ],
